보스 달리기 패턴 수정
보스 그로기 이펙트 수정
플레이어 골드 기능 추가
골드 ui 추가
BossScene light rendering 수정
GameScene light rendering 수정
플레이어 사망시 보스 patrol

+++ Assets/Scripts/Shop.cs
@@ -7,7 +7,7 @@ using UnityEngine.Serialization;
 public class Shop : MonoBehaviour
 {
     [SerializeField] private ShopStand[] itemStands;
-    [SerializeField] private int priceUnit;
+    [SerializeField] private float priceUnit;
 
     private void Start()
     {
